Adds OCIO scraping to Nuke. Extracts OCIO scraping from Maya utils into it's own file
